feat: add yaml-has-properties asserter
