+++ package.json
@@ -13,13 +13,14 @@
     "@octokit/rest": "^16.34.1",
     "cli-ux": "^5.3.3",
     "fs-extra": "^8.1.0",
-    "js-yaml": "^3.13.1"
+    "js-yaml": "^3.14.1"
   },
   "devDependencies": {
     "@oclif/dev-cli": "^1",
     "@oclif/test": "^1",
     "@types/chai": "^4",
     "@types/fs-extra": "^8.0.1",
+    "@types/js-yaml": "^3.12.6",
     "@types/mocha": "^5",
     "@types/node": "^10",
     "chai": "^4",
