+++ src/asserters/index.ts
@@ -1,5 +1,6 @@
 import {FileExactMatchAsserter, FileDoesNotExistAsserter} from './file'
 import {JsonHasPropertiesAsserter} from './json'
+import {YamlHasPropertiesAsserter} from './yaml'
 import {NodeProjectHasDepsAsserter, NodeProjectDoesNotHaveDepsAsserter} from './node'
 
 export const AsserterLookup: { [key: string]: any } = {
@@ -8,4 +9,5 @@ export const AsserterLookup: { [key: string]: any } = {
   'json-has-properties': JsonHasPropertiesAsserter,
   'node-project-has-deps': NodeProjectHasDepsAsserter,
   'node-project-does-not-have-deps': NodeProjectDoesNotHaveDepsAsserter,
+  'yaml-has-properties': YamlHasPropertiesAsserter,
 }
